Move RestoreWorkerInterface into fdbserver
