+++ fdbclient/SystemData.h
@@ -26,7 +26,6 @@
 
 #include "fdbclient/FDBTypes.h"
 #include "fdbclient/StorageServerInterface.h"
-#include "fdbclient/RestoreWorkerInterface.actor.h"
 
 // Don't warn on constants being defined in this file.
 #pragma clang diagnostic push
@@ -443,31 +442,6 @@ extern const KeyRef mustContainSystemMutationsKey;
 
 // Key range reserved for storing changes to monitor conf files
 extern const KeyRangeRef monitorConfKeys;
-
-// Fast restore
-extern const KeyRef restoreLeaderKey;
-extern const KeyRangeRef restoreWorkersKeys;
-extern const KeyRef restoreStatusKey; // To be used when we measure fast restore performance
-extern const KeyRef restoreRequestTriggerKey;
-extern const KeyRef restoreRequestDoneKey;
-extern const KeyRangeRef restoreRequestKeys;
-extern const KeyRangeRef restoreApplierKeys;
-extern const KeyRef restoreApplierTxnValue;
-
-const Key restoreApplierKeyFor(UID const& applierID, int64_t batchIndex, Version version);
-std::tuple<UID, int64_t, Version> decodeRestoreApplierKey(ValueRef const& key);
-const Key restoreWorkerKeyFor(UID const& workerID);
-const Value restoreWorkerInterfaceValue(RestoreWorkerInterface const& server);
-RestoreWorkerInterface decodeRestoreWorkerInterfaceValue(ValueRef const& value);
-const Value restoreRequestTriggerValue(UID randomUID, int const numRequests);
-int decodeRestoreRequestTriggerValue(ValueRef const& value);
-const Value restoreRequestDoneVersionValue(Version readVersion);
-Version decodeRestoreRequestDoneVersionValue(ValueRef const& value);
-const Key restoreRequestKeyFor(int const& index);
-const Value restoreRequestValue(RestoreRequest const& server);
-RestoreRequest decodeRestoreRequestValue(ValueRef const& value);
-const Key restoreStatusKeyFor(StringRef statusType);
-const Value restoreStatusValue(double val);
 
 extern const KeyRef healthyZoneKey;
 extern const StringRef ignoreSSFailuresZoneString;
