+++ fdbserver/RestoreCommon.actor.h
@@ -35,6 +35,7 @@
 #include "fdbclient/NativeAPI.actor.h"
 #include "fdbrpc/IAsyncFile.h"
 #include "fdbclient/BackupAgent.actor.h"
+#include "fdbserver/Knobs.h"
 
 #include "flow/actorcompiler.h" // has to be last include
