+++ fdbserver/RestoreWorker.actor.h
@@ -33,12 +33,12 @@
 #include <cstdint>
 #include <cstdarg>
 
-#include "fdbclient/RestoreWorkerInterface.actor.h"
 #include "fdbserver/RestoreUtil.h"
 #include "fdbserver/RestoreCommon.actor.h"
 #include "fdbserver/RestoreRoleCommon.actor.h"
 #include "fdbserver/RestoreLoader.actor.h"
 #include "fdbserver/RestoreApplier.actor.h"
+#include "fdbserver/RestoreWorkerInterface.actor.h"
 
 // Each restore worker (a process) is assigned for a role.
 // MAYBE Later: We will support multiple restore roles on a worker
